% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_function.r
\name{complex_function}
\alias{complex_function}
\title{Complex function that does some task}
\usage{
complex_function(opt, k)
}
\arguments{
\item{opt, }{description}

\item{k, }{description}
}
\description{
Complex function that does some task
}
