% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_custom_template}
\alias{create_custom_template}
\title{Create a template}
\usage{
create_custom_template(input)
}
\arguments{
\item{input}{input$template_name_input (template name)
input$template_desc_input (template description)
input$rmd_input (template body)}
}
\description{
Create a user-defined template, saved in the
templater package folder.
}
\examples{
\dontrun{
if(interactive()) {
 input <- NULL
 input$template_name_input <- "Title"
 input$template_desc_input <- "A description"
 input$rmd_input <- "Lorem ipsum"
 create_custom_template(input)
 }
}
}
\seealso{
\code{\link[fs]{create}},
\code{\link[fs]{file_access}}
\code{\link[rstudioapi]{isAvailable}},
\code{\link[rstudioapi]{showDialog}}
}
