% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_other_templates}
\alias{get_other_templates}
\title{get other templates}
\usage{
get_other_templates()
}
\value{
tibble
}
\description{
Find all templater templates (including custom templates)
stored locally on the computer, and return a dataframe for later use.
}
\examples{
\dontrun{
if(interactive()) {
 get_other_templates()
 }
}
}
