% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_writeable_lib}
\alias{get_writeable_lib}
\title{Get Writeable Library}
\usage{
get_writeable_lib()
}
\value{
file path
}
\description{
Returns the first writeable R library found.
}
\examples{
\dontrun{
if(interactive()) {
  get_writeable_lib()
 }
}
}
\seealso{
\code{\link[fs]{file_access}}
}
