% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{Check if Path is Absolute}
\usage{
is_absolute_path(x)
}
\arguments{
\item{x}{The paths to be checked}
}
\value{
A logical vector.
}
\description{
Absolute paths start with either ~ or / (linux) or X:/ or X:\\ (windows).
}
