% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_mult_c.r
\name{mat_mult_c}
\alias{mat_mult_c}
\title{Function to multiply two matrices in C (iterative algorithm)}
\usage{
mat_mult_c(A, B)
}
\arguments{
\item{a}{Matrix a}

\item{b}{Matrix b}
}
\description{
Contains a triple for loop to multiply two matrices implemented in C.
Warning: this function is just for pedagogical purposes. Don't use this
function for computation
}
\examples{
A <- matrix(rnorm(10000), 100, 100)
B <- matrix(rnorm(10000), 100, 100)
system.time(AB <- mat_mult_c(A,B))
}
\seealso{
mat_mult_f mat_mult
}
\author{
Jordan Schupbach
}
\keyword{Matrix}
\keyword{Multiplication}
