% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_template}
\alias{render_template}
\title{Safely Compile a R/markdown Template}
\usage{
render_template(template, output = gsub(".Rmd$", ".html", template),
  html = grepl("html$", output), md.file = gsub("[.](html|md)$", ".md",
  output), resource.dir = system.file(package = "templater"),
  macros = "macros.tex", opts.knit = NULL)
}
\arguments{
\item{template}{Name of .Rmd file.}

\item{output}{Name of output .md or .html file.}

\item{html}{Run pandoc to compile html from the .md?}

\item{md.file}{Name of the .md output file.}

\item{resource.dir}{Directory that files referenced in pandoc options are in.}

\item{macros}{Name of a file with LaTeX macros.}
}
\value{
The name of the output file.
Note that by default, both knitr and pandoc look for figures relative to the *current directory*,
not the directory that the markdown file lives in.  
This makes subsequent parsing of the markdown file difficult.
To avoid this, here we switch to the directory of the markdown file before running either.

If the name of the output is 'outdir/out.html', then the 'cache' and 'figure' directories
for knitr will be 'outdir/cache/out/' and 'outdir/figure/out', respectively.
This ensures that separate output files have distinct caches and figures.
}
\description{
Calls "knit()" on the template given,
directing output to the given output name,
and doing this in a way that doesn't cause collisions 
between parallel instances of this script.
}

