% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{Source all function.}
\usage{
source_all(path = "R/", pattern = "*.R$", packages = "dplyr")
}
\description{
Source all R files in a directory at once. 
Also loads selected libraries. 
This function can be useful when developing R packages, 
instead of rebuilding the entire package every time you make a change 
(which is a more robust way to checking the package works, 
but is much slower).
}
\keyword{internal}
