% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater.R
\name{templater}
\alias{templater}
\title{Create a standardized project directory}
\usage{
templater(project_name, readme = TRUE)
}
\arguments{
\item{project_name}{The name of the project (and the root folder)}

\item{readme}{Logical; should a README be generated?}
}
\description{
Create a standardized project directory
}
