% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templater.R
\name{use_templater}
\alias{use_templater}
\title{Use templater}
\usage{
use_templater()
}
\value{
None
}
\description{
This function uses the usethis package to create a folder for R functions, a
folder for data, a folder for documents, a file for targets to orchestrate
the project workflow, an .Renviron file, and an .Rprofile file.
}
\examples{
use_templater()
}
