% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{write_table_cache}
\alias{write_table_cache}
\title{Write Out a Table to knitr's Cache Directory}
\usage{
write_table_cache(x, file = paste(knitr::opts_current$get("label"), ".csv",
  sep = ""), ...)
}
\arguments{
\item{x}{Data to write out.}

\item{file}{File name to write to.}

\item{...}{Other parameters passed to write.csv().}
}
\value{
The name of the output file.
}
\description{
Writes out the given table to a text file to the currently used cache directory,
and returns the path to the file written.
}

