% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_package_templates}
\alias{get_package_templates}
\title{get package templates}
\usage{
get_package_templates()
}
\value{
tibble
}
\description{
Find all package templates stored locally on the computer,
return a dataframe for later use.
}
\examples{
\dontrun{
if(interactive()) {
 get_package_templates()
 }
}
}
