% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{Source all functions}
\usage{
source_all(path = "R/", pattern = "*.R$", packages = "dplyr", ...)
}
\arguments{
\item{path}{Directory containing R files.}

\item{pattern}{File name pattern to search for.}

\item{packages}{Packages to install and load.}

\item{...}{Additional arguments passed to \link[base]{source}.}
}
\value{
Null
}
\description{
Source all R files in a directory at once. 
Also loads selected libraries. 
This function can be useful when developing R packages, 
instead of rebuilding the entire package every time you make a change 
(which is a more robust way to checking the package works, 
but is much slower).
}
\keyword{internal}
