% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater.R
\name{templater}
\alias{templater}
\title{Call templater}
\usage{
templater()
}
\description{
Opens the templater shiny gadget, allowing
for template usage and creation.
}
\examples{
\dontrun{
if(interactive()) {
   templater()
 }
}
}
\seealso{
\code{\link[shiny]{runGadget}},
\code{\link[shiny]{shinyApp}},
\code{\link[shiny]{viewer}}
}
