% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater.r
\name{TemplateR}
\alias{TemplateR}
\title{A simple template to generate R packages}
\usage{
TemplateR(opt = "things", k = 0.5)
}
\arguments{
\item{opt}{A string, specifying some function parameter. Default is \code{"things"}}

\item{k}{A double between 0 and 1. Description of what it does.
Default is \code{0.5}.}
}
\value{
A string.
}
\description{
This should be the main function of your package and this should
be its description.
}
\examples{
library("TemplateR")

# run function
templater_res <- TemplateR(opt="things", k=0.2)

# check results
templater_res

}
